% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_plot.R
\name{draw_key_cust}
\alias{draw_key_cust}
\title{Draw custom legend key for fpci_plot}
\usage{
draw_key_cust(data, params, size)
}
\arguments{
\item{data}{data}

\item{params}{params}

\item{size}{size}
}
\value{
key glyph for legend
}
\description{
Draw custom legend key for fpci_plot
}
