% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_plot.R
\name{st_plot_treated}
\alias{st_plot_treated}
\title{Make Science Table Plot with treatment assignments shown}
\usage{
st_plot_treated(data, color = TRUE, title = "", size = 0.8)
}
\arguments{
\item{data}{input data with y1 and y0 shown}

\item{color}{color dots based on treatment effect (positive, negative, zero)}

\item{title}{plot title}

\item{size}{dot size for geom_point}
}
\value{
ggplot object
}
\description{
Produces a science table plot from input data
}
\examples{
pos_effect <- data.frame(y0 = rnorm(n = n, 2.5)) \%>\%
    mutate(y1 = y0  +  1.25 + rnorm(n = n, sd = 0.1))
st_plot(pos_effect, title = "Positive effect")
}
