% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_plot.R
\name{superpop_plot}
\alias{superpop_plot}
\title{Superpopulation Science table plot}
\usage{
superpop_plot(data, superpop_data, title = "", size = 0.8)
}
\arguments{
\item{data}{input data with y1 and y0 shown}

\item{superpop_data}{like data, but with many more observations (representing
a superpopulation)}

\item{title}{plot title}

\item{size}{dot size for geom_point}
}
\value{
ggplot object
}
\description{
Superpopulation Science table plot
}
